requirements-dev.txt
-r requirements.txt
django-nose==1.4.6
ipdb==0.12
